Input system changed.
Changed from the old system to the new system. Warning, InputAction "DuckTape" required in PlayerInput, else causes a null reference exception. Whoever is working on the action mappings, let me know the name of the actual "DuckTape" button and it should be fixed after a merge.

That is the only problem this code faces.

So basically fix up the PlayerController & PlayerInput stuff and it should be good to go.