Scores added (Fixed value)
The original intent was that each game could have it's own point value. Instead, due to an unknown bug, the score is added by 100. The bug is marked in the code and seems to stem from the AMiniGame parent class.
The bug has been bypassed and increasing the player score works, but not as intended.